Replace `reqwest` with `bitreq`
`reqwest` has an absolutely insane dependency tree, with
anti-features (like insanely-fast-unsafe-IDN parsing) baked in.
Instead, the rust-bitcoin ecosystem is starting to move to `bitreq`
which is relatively simple, only depends (even optionally) on
`tokio` and `rustls`, and doesn't bother with crap like IDNs.

Here we switch over, maintaining the ability to proxy connections
for those that need that.